Fix react-refresh warning in public-chat page
Move getPublicChatAccessErrorMessage into chatkit/lib so the page
module only exports components.
